Add for fun level text changes.
This allows for changing the title text of the levels to be something other than 'LEVEL-'. Eventually, it would be nice to turn this into either a flexible string attribute or allow for more choices.

A character formatter class has also been put in a little early to allow for eventual text edits of the various NPCs that ramble.

All of the text based classes have been moved to their own sub-package to better keep track of them. I am aware that the constants file may have been the best place for them before, but considering how much we can change here, we should consider splitting things up.