Use settings and flags objects.
This sets up settings that can be passed along to the randomizer. The flags are a part of the settings.

An attempt was made to allow for random level text by using `*`. It will pick from a predefined list. Any valid name can be supplied, however.

I am slightly concerned about a few things with this PR, though.

* Lately, any attempts at randomizing the text speed always seems to go to very slow: aka, 10 frames per letter. That is not exactly good.
* Is there a standard way of updating the GUI code? The comments seem to indicate something was used, and if I were to make a change, it will risk being lost.

Hopefully we can figure out what is going on.